requirements.txt
# ==========================================
# Project: Analog Knowledge Base
# Dependencies for Web Scraping and Local RAG
# ==========================================

# 1. Web Scraping & HTML Parsing
requests
beautifulsoup4
markdownify

# 2. Vector Database & RAG Framework
chromadb
langchain
langchain-community
langchain-google-genai

# 3. Environment & Utilities
python-dotenv